requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic[GetJsonSchemaHandle]
passlib[bcrypt]
openai
pytest
httpx
pymongo
motor
python-docx